+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 ## Next release
 
 - feat : l3 support
+- fix(block_production): dynamic block closing now adds special address with prev block hash
+- fix(rpc): call, simulate, estimate rpcs executed on top of the block, not at the start of it
+- fix(compilation): crate-level compilation
+- chore: Move crates under a madara subdir
+- chore(nix): resolve flake and direnv compatibility issues
+- fix: Gateway path fix
+- fix: instrumentation code
+- feat: block resource cap removed from the pending tick
+- fix: replace class hash issue resolved + gas fees issue resolved
+- fix: trim hash of eth state was failing with 0x0
 - fix: devnet accounts getting deployed in sequencer mode
 - fix(rpc): fix BroadcastedDeclareTxn V3 in starknet-types-rpc
 - fix: oracle need condition